refactor: move utils/ to project root
