+++ utils/getStatusBar.js
